moved Form to Form.jsx, added: test files
